LAB10.15_Deleting Files

+++ controllers/admin.js
@@ -2,6 +2,7 @@ const { validationResult } = require('express-validator');
 const mongoose = require('mongoose');
 
 const Product = require("../models/product");
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -172,6 +173,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then(result => {
@@ -188,7 +190,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodID = req.body.productID;
-  Product.deleteOne({ _id: prodID, userId: req.user._id })
+  Product.findById(prodID)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodID, userId: req.user._id });
+    })
     .then(() => {
       console.log('DELETED PRODUCT');
       res.redirect('/admin/products');
